chore: move navigation and footer into feature namespace
